shared dtype uses
