+++ pantab/_writermodule.c
@@ -3,6 +3,7 @@
 #include <datetime.h>
 #include <inttypes.h>
 
+#include "dtypes.h"
 #include "tableauhyperapi.h"
 
 #define MICROSECONDS_PER_DAY 24 * 60 * 60 * 1000000
@@ -14,75 +15,6 @@ typedef struct {
     int32_t months;
 } py_interval;
 
-typedef enum {
-    INT16_ = 1,
-    INT32_,
-    INT64_,
-    INT16NA = 6,
-    INT32NA,
-    INT64NA,
-    FLOAT32_ = 11,
-    FLOAT64_,
-    BOOLEAN = 50,
-    DATETIME64_NS = 100,
-    DATETIME64_NS_UTC,
-    TIMEDELTA64_NS = 200,
-    OBJECT = 220,
-    UNKNOWN = 255
-} DTYPE;
-
-const static struct {
-    DTYPE dtype;
-    const char *str;
-} dtype_map[] = {{INT16_, "int16"},
-                 {INT32_, "int32"},
-                 {INT64_, "int64"},
-                 {INT16NA, "Int16"},
-                 {INT32NA, "Int32"},
-                 {INT64NA, "Int64"},
-                 {FLOAT32_, "float32"},
-                 {FLOAT64_, "float64"},
-                 {BOOLEAN, "bool"},
-                 {DATETIME64_NS, "datetime64[ns]"},
-                 {DATETIME64_NS_UTC, "datetime64[ns, UTC]"},
-                 {TIMEDELTA64_NS, "timedelta64[ns]"},
-                 {OBJECT, "object"}};
-
-static DTYPE stringToDtype(const char *str) {
-    for (Py_ssize_t i = 0;
-         i < (Py_ssize_t)(sizeof(dtype_map) / sizeof(dtype_map[0])); i++) {
-        if (strcmp(str, dtype_map[i].str) == 0) {
-            return dtype_map[i].dtype;
-        }
-    }
-
-    return UNKNOWN;
-}
-
-// Caller is responsible for returned object
-static DTYPE *makeEnumeratedDtypes(PyTupleObject *obj) {
-    Py_ssize_t len = PyTuple_GET_SIZE(obj);
-    DTYPE *result = malloc(len * sizeof(DTYPE));
-
-    for (Py_ssize_t i = 0; i < len; i++) {
-        PyObject *dtypeObj = PyTuple_GET_ITEM(obj, i);
-        const char *dtypeStr = PyUnicode_AsUTF8(dtypeObj);
-        DTYPE dtype = stringToDtype(dtypeStr);
-
-        if (dtype == UNKNOWN) {
-            free(result);
-            PyObject *errMsg =
-                PyUnicode_FromFormat("Unknown dtype: \"%s\"\n", dtypeStr);
-            PyErr_SetObject(PyExc_TypeError, errMsg);
-            Py_DECREF(errMsg);
-            return NULL;
-        }
-
-        result[i] = dtype;
-    }
-
-    return result;
-}
 
 static int isNull(PyObject *data) {
     if ((data == Py_None) ||
